+++ src/message/GetMessageTextHtmlRequest.ts
@@ -1,0 +1,26 @@
+import { Request } from '../Request';
+import { IRequestOptions, IRestResponse } from 'typed-rest-client/RestClient';
+import restClient from '../MailinatorRestClient';
+import { AUTHORIZATION } from '../Constants';
+
+const _resolveTemplateUrl = (domain: string, messageId: string) => {
+    return `https://api.mailinator.com/api/v2/domains/${domain}/messages/${messageId}/texthtml`;
+};
+
+export class GetMessageTextHtmlRequest implements Request<Record<string, unknown>> {
+
+    constructor(private readonly domain: string,
+                private readonly messageId: string) {
+    }
+
+    execute(apiToken: string): Promise<IRestResponse<Record<string, unknown>>> {
+
+        const _options: IRequestOptions = {
+            additionalHeaders: {
+                [AUTHORIZATION]: apiToken
+            }
+        };
+
+        return restClient.get<Record<string, unknown>>(_resolveTemplateUrl(this.domain, this.messageId), _options);
+    }
+}