+++ src/message/MessageHeadersResponse.ts
@@ -1,0 +1,3 @@
+export class MessageHeadersResponse {
+    headers!: Record<string, unknown>;
+}